requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
anthropic>=0.79.0
PyPDF2>=3.0.0
python-docx>=1.0.0
gunicorn>=21.0.0
